# api/openapi.yaml
openapi: 3.1.0
info:
  title: Ydisks 闲鱼助手 API
  version: 1.0.0
  description: 版本化 HTTP API 的唯一契约源。
servers:
  - url: /
paths:
...
  "/api/v1/analytics/orders":
    get:
      operationId: getApiV1AnalyticsOrders
      summary: getApiV1AnalyticsOrders
      security: [{cookieAuth: []}]
      parameters:
        - name: start_date
          in: query
          required: true
          schema: {type: string, format: date}
        - name: end_date
          in: query
          required: true
          schema: {type: string, format: date}
        - name: timezone_offset_minutes
          in: query
          required: false
          schema: {type: integer}
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnalyticsResponse'
        '400':
          description: 统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: 统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: 统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: 统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  "/api/v1/analytics/orders/valid":
    get:
      operationId: getApiV1AnalyticsOrdersValid
      summary: getApiV1AnalyticsOrdersValid
      security: [{cookieAuth: []}]
      parameters:
        - name: start_date
          in: query
          required: true
          schema: {type: string, format: date}
        - name: end_date
          in: query
          required: true
          schema: {type: string, format: date}
        - name: timezone_offset_minutes
          in: query
          required: false
          schema: {type: integer}
        - name: page
          in: query
          required: false
          schema: {type: integer, minimum: 1}
        - name: page_size
          in: query
          required: false
          schema: {type: integer, minimum: 1, maximum: 500}
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidOrdersResponse'
        '400':
          description: 统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: 统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: 统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: 统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  "/api/v1/admin/data-backup":
    get:
      operationId: downloadDataBackup
      summary: 下载当前 SQLite 数据库一致性备份
      security: [{cookieAuth: []}]
      x-contract-success-exception:
        kind: binary
      responses:
        '200':
          description: SQLite 数据库备份文件
          headers:
            Content-Disposition: {schema: {type: string}}
            Cache-Control: {schema: {type: string}}
          content:
            application/vnd.sqlite3:
              schema: {type: string, format: binary}
        '400': {description: 统一错误响应, content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}}
        '401': {description: 统一错误响应, content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}}
        '403': {description: 统一错误响应, content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}}
        '422': {description: 当前数据库不支持内置备份, content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}}
        '500': {description: 统一错误响应, content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}}
  "/api/v1/admin/data-restore":
    post:
      operationId: importDataBackup
      summary: 校验并暂存 SQLite 数据库恢复文件
      security: [{cookieAuth: []}]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file, confirmation]
              properties:
                file: {type: string, format: binary}
                confirmation: {type: string, enum: [RESTORE]}
      responses:
        '200': {description: 恢复文件已暂存, content: {application/json: {schema: {$ref: '#/components/schemas/DataRestoreResponse'}}}}
        '400': {description: 统一错误响应, content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}}
        '401': {description: 统一错误响应, content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}}
        '403': {description: 统一错误响应, content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}}
        '422': {description: 当前数据库不支持内置恢复, content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}}
        '500': {description: 统一错误响应, content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}}
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: session
  schemas:
    DataRestoreResponse:
      type: object
      required: [filename, size, restart_required, message]
      properties:
        filename: {type: string}
        size: {type: integer, format: int64, minimum: 1}
        restart_required: {type: boolean}
        message: {type: string}
...
    OrderDTO:
      type: object
      required: [order_id, item_id, item_title, item_image, buyer_id, buyer_name, quantity, amount, order_status, status, cookie_id, created_at]
      properties:
        order_id: {type: string}
        item_id: {type: string}
        item_title: {type: string}
        item_image: {type: string}
        buyer_id: {type: string}
        buyer_name: {type: string}
        spec_name: {type: string}
        spec_value: {type: string}
        quantity: {type: string}
        amount: {type: string}
        order_status: {type: string}
        status: {type: string}
        cookie_id: {type: string}
        is_bargain: {type: integer}
        system_shipped: {type: boolean}
        receiver_name: {type: string}
        receiver_phone: {type: string}
        receiver_address: {type: string}
        receiver_city: {type: string}
        created_at: {type: string}
        updated_at: {type: string}
...
    AnalyticsResponse:
      type: object
      required: [revenue_stats, daily_stats, status_stats, city_stats, item_stats]
      properties:
        revenue_stats: {type: object, required: [total_orders, total_amount, avg_amount, unique_buyers, unique_items], properties: {total_orders: {type: integer}, total_amount: {type: number}, avg_amount: {type: number}, unique_buyers: {type: integer}, unique_items: {type: integer}}}
        daily_stats: {type: array, items: {type: object, required: [date, order_count, amount], properties: {date: {type: string}, order_count: {type: integer}, amount: {type: number}}}}
        status_stats: {type: array, items: {type: object, required: [status, count, amount], properties: {status: {type: string}, count: {type: integer}, amount: {type: number}}}}
        city_stats: {type: array, items: {type: object, required: [city, order_count, total_amount], properties: {city: {type: string}, order_count: {type: integer}, total_amount: {type: number}}}}
        item_stats: {type: array, items: {type: object, required: [item_id, order_count, total_amount, avg_amount], properties: {item_id: {type: string}, order_count: {type: integer}, total_amount: {type: number}, avg_amount: {type: number}}}}
    ValidOrdersResponse:
      type: object
      required: [orders, total, page, page_size, truncated]
      properties:
        orders: {type: array, items: {$ref: '#/components/schemas/OrderDTO'}}
        total: {type: integer}
        page: {type: integer}
        page_size: {type: integer}
        truncated: {type: boolean}
...
    ErrorResponse:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string
        request_id:
          type: string
        details:
          type: object
      additionalProperties: true